OpenStack Cinder CSI volume provisioner (#15)
This guide is designed to help you effortlessly integrate the Cinder CSI Volume Provisioner into your OKD or OpenShift cluster. This guide is especially valuable for those who have configured their cluster with the platform option set to "none" and, as a result, are missing the OpenStack Cinder CSI Driver Operator.